requirements.txt
requests>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
ipython>= 9.4.0